+++ src/third-party/singleapplication/examples/basic/main.cpp
@@ -1,9 +1,0 @@
-#include <SingleApplication.h>
-
-int main(int argc, char *argv[])
-{
-    // Allow secondary instances
-    SingleApplication app( argc, argv );
-
-    return app.exec();
-}